Move FileSystemTestUtils to test
