+++ tests/src/test/java/alluxio/client/IsolatedFileSystemIntegrationTest.java
@@ -17,6 +17,7 @@ import alluxio.LocalAlluxioClusterResource;
 import alluxio.PropertyKey;
 import alluxio.client.file.FileInStream;
 import alluxio.client.file.FileSystem;
+import alluxio.client.file.FileSystemTestUtils;
 import alluxio.client.file.URIStatus;
 import alluxio.client.file.options.CreateFileOptions;
 import alluxio.heartbeat.HeartbeatContext;
