+++ tests/src/test/java/alluxio/client/UnderStorageReadIntegrationTest.java
@@ -18,6 +18,7 @@ import alluxio.PropertyKey;
 import alluxio.client.file.FileInStream;
 import alluxio.client.file.FileOutStream;
 import alluxio.client.file.FileSystem;
+import alluxio.client.file.FileSystemTestUtils;
 import alluxio.client.file.options.CreateFileOptions;
 import alluxio.client.file.options.OpenFileOptions;
 import alluxio.util.io.BufferUtils;
