+++ tests/src/test/java/alluxio/hadoop/HdfsFileInputStreamIntegrationTest.java
@@ -14,11 +14,11 @@ package alluxio.hadoop;
 import alluxio.AlluxioURI;
 import alluxio.LocalAlluxioClusterResource;
 import alluxio.PropertyKey;
-import alluxio.client.FileSystemTestUtils;
 import alluxio.client.ReadType;
 import alluxio.client.WriteType;
 import alluxio.client.file.FileSystem;
 import alluxio.client.file.FileSystemContext;
+import alluxio.client.file.FileSystemTestUtils;
 import alluxio.client.file.URIStatus;
 import alluxio.client.util.ClientTestUtils;
 import alluxio.exception.AlluxioException;
