+++ tests/src/test/java/alluxio/shell/command/CopyToLocalCommandTest.java
@@ -12,8 +12,8 @@
 package alluxio.shell.command;
 
 import alluxio.SystemPropertyRule;
-import alluxio.client.FileSystemTestUtils;
 import alluxio.client.WriteType;
+import alluxio.client.file.FileSystemTestUtils;
 import alluxio.shell.AbstractAlluxioShellTest;
 import alluxio.shell.AlluxioShellUtilsTest;
