+++ tests/src/test/java/alluxio/shell/command/FreeCommandTest.java
@@ -13,8 +13,8 @@ package alluxio.shell.command;
 
 import alluxio.AlluxioURI;
 import alluxio.IntegrationTestUtils;
-import alluxio.client.FileSystemTestUtils;
 import alluxio.client.WriteType;
+import alluxio.client.file.FileSystemTestUtils;
 import alluxio.exception.AlluxioException;
 import alluxio.heartbeat.HeartbeatContext;
 import alluxio.heartbeat.ManuallyScheduleHeartbeat;
